requirements.txt
Django==2.1.4
djangorestframework==3.9.0
dj-static==0.0.6
dj-database-url==0.4.2
gunicorn==19.7.1
pytz==2017.3
whitenoise==3.3.1
Unipath==1.0
python-decouple==3
Pillow==3.3.0
Markdown==2.6.6
bleach==1.4.3
psycopg2==2.7.5
django-filter==2.1.0


